MASSIVE CHANGES EVERYTHING BROKEN
Starting work on decoupling the Session and Message code
Adding StartTLS functionality
Updated Launcher News/Updates code to match updated website contract

NOTE: This update is part of a work in progress that I allowed to grow
too big. It breaks basically everything even though it builds
successfully. A lot of work still needs to be done to finish the
project.